Create dbmock package and add mocks with Options
