+++ cmd/frontend/graphqlbackend/access_requests_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/auth"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/accessrequests"
+	"github.com/sourcegraph/sourcegraph/internal/database/dbmock"
 	"github.com/sourcegraph/sourcegraph/internal/types"
 )
 
@@ -33,7 +34,7 @@ func TestAccessRequestNode(t *testing.T) {
 	accessRequestStore := accessrequests.NewMockStore()
 	accessRequestStore.GetByIDFunc.SetDefaultReturn(mockAccessRequest, nil)
 
-	db := database.Mock(mockDB, accessRequestStore)
+	db := dbmock.NewMockedDB(mockDB, dbmock.WithMock(accessRequestStore))
 
 	ctx := actor.WithActor(context.Background(), &actor.Actor{UID: 1})
 
@@ -89,7 +90,7 @@ func TestAccessRequestsQuery(t *testing.T) {
 	mockDB.UsersFunc.SetDefaultReturn(userStore)
 
 	accessRequestStore := accessrequests.NewMockStore()
-	db := database.Mock(mockDB, accessRequestStore)
+	db := dbmock.NewMockedDB(mockDB, dbmock.WithMock(accessRequestStore))
 
 	t.Parallel()
 
@@ -194,7 +195,7 @@ func TestSetAccessRequestStatusMutation(t *testing.T) {
 
 	t.Run("non-admin user", func(t *testing.T) {
 		accessRequestStore := accessrequests.NewMockStore()
-		db := database.Mock(mockDB, accessRequestStore)
+		db := dbmock.NewMockedDB(mockDB, dbmock.WithMock(accessRequestStore))
 
 		userStore.GetByCurrentAuthUserFunc.SetDefaultReturn(&types.User{ID: 1, SiteAdmin: false}, nil)
 		ctx := actor.WithActor(context.Background(), &actor.Actor{UID: 1})
@@ -221,7 +222,7 @@ func TestSetAccessRequestStatusMutation(t *testing.T) {
 
 	t.Run("existing access request", func(t *testing.T) {
 		accessRequestStore := accessrequests.NewMockStore()
-		db := database.Mock(mockDB, accessRequestStore)
+		db := dbmock.NewMockedDB(mockDB, dbmock.WithMock(accessRequestStore))
 
 		createdAtTime, _ := time.Parse(time.RFC3339, "2023-02-24T14:48:30Z")
 		mockAccessRequest := &types.AccessRequest{ID: 1, Email: "a1@example.com", Name: "a1", CreatedAt: createdAtTime, AdditionalInfo: "af1", Status: types.AccessRequestStatusPending}
@@ -248,7 +249,7 @@ func TestSetAccessRequestStatusMutation(t *testing.T) {
 
 	t.Run("non-existing access request", func(t *testing.T) {
 		accessRequestStore := accessrequests.NewMockStore()
-		db := database.Mock(mockDB, accessRequestStore)
+		db := dbmock.NewMockedDB(mockDB, dbmock.WithMock(accessRequestStore))
 
 		notFoundErr := &accessrequests.ErrAccessRequestNotFound{ID: 1}
 		accessRequestStore.GetByIDFunc.SetDefaultReturn(nil, notFoundErr)
