+++ internal/database/database.go
@@ -3,7 +3,6 @@ package database
 import (
 	"context"
 	"database/sql"
-	"reflect"
 	"time"
 
 	"github.com/sourcegraph/log"
@@ -102,35 +101,6 @@ func NewDBWith(logger log.Logger, other basestore.ShareableStore) DB {
 type db struct {
 	*basestore.Store
 	logger log.Logger
-}
-
-type mockedDB struct {
-	DB
-	mockedStore reflect.Value
-}
-
-func (mdb *mockedDB) WithTransact(ctx context.Context, f func(tx DB) error) error {
-	return mdb.DB.WithTransact(ctx, func(tx DB) error {
-		return f(&mockedDB{DB: tx, mockedStore: mdb.mockedStore})
-	})
-}
-
-func GetMock[T basestore.ShareableStore](db DB) (t T) {
-	switch v := db.(type) {
-	case *mockedDB:
-		if v.mockedStore.Type().Implements(reflect.TypeOf((*T)(nil)).Elem()) {
-			return v.mockedStore.Interface().(T)
-		}
-		return GetMock[T](v.DB)
-	}
-	return t
-}
-
-func Mock[T basestore.ShareableStore](db DB, val T) *mockedDB {
-	return &mockedDB{
-		DB:          db,
-		mockedStore: reflect.ValueOf(val),
-	}
 }
 
 func (d *db) Logger() log.Logger {
